ai_principal/rag_preprocessing/requirements.txt
openai>=1.1.0
numpy>=1.26.0
supabase>=1.2.0
python-dotenv==1.0.0
tiktoken>=0.5.2